Add error handling for apis calls on users show
The users show page ("settings") relies on access to Unsplash data, but
previous when the api was unavailable, the page would error out.

With these unpdates instead of an error, a flash message is placed on
the page stating that Unsplash is unavailable.

This is handled in the UnsplashSettings class with an error instance
variable.  All of the class's api calls have error handling that will
set @error to true if the api returns any errors.  The presence of an
error is checked first before any api calls to prevent from hitting the
api on the same request when an error has already been found.